Support packing/unpacking of custom slices
Allows custom types containing slices to handle their own elem
processing. Does not affect slices of a custom type.

Useful for implementing e.g. NULL-terminated slices whose length
is unknown and cannot be determined by a field in the struct.

    type IntSlice []int
    type MyInt int
    type Fields struct {
        intsA IntSlice // Handles own, fd.Slice = false
        intsB []MyInt  // Doesn't handle own, fd.Slice = true
    }

This is one method to inform the parsing process of the different
requirements. I'm not sure it is ideal. Another option is new
method, such as `HandlesElements()` on the Custom interface.

Closes #60

Co-authored-by: Robin Eklind <mewmew@users.noreply.github.com>